ContentPersistenceService -> ObjectStorageService
