+++ src/main/java/org/ambraproject/rhino/service/impl/ArticleDatabaseServiceImpl.java
@@ -36,7 +36,7 @@ import org.ambraproject.rhino.model.ingest.ArticlePackage;
 import org.ambraproject.rhino.model.ingest.IngestPackage;
 import org.ambraproject.rhino.rest.RestClientException;
 import org.ambraproject.rhino.service.ConfigurationReadService;
-import org.ambraproject.rhino.service.ContentPersistenceService;
+import org.ambraproject.rhino.service.ObjectStorageService;
 import org.ambraproject.rhino.service.ArticleDatabaseService;
 import org.ambraproject.rhino.service.JournalCrudService;
 import org.hibernate.Query;
@@ -58,7 +58,7 @@ public class ArticleDatabaseServiceImpl implements ArticleDatabaseService {
   @Autowired
   private JournalCrudService journalCrudService;
   @Autowired
-  private ContentPersistenceService contentRepoPersistenceService;
+  private ObjectStorageService objectStorageService;
   @Autowired
   private ConfigurationReadService configurationReadService;
 
@@ -115,13 +115,13 @@ public class ArticleDatabaseServiceImpl implements ArticleDatabaseService {
   @Override
   public void persistAssets(ArticlePackage articlePackage, ArticleIngestion ingestion) {
     List<ArticleItem> items = articlePackage.getAllItems().stream()
-        .map((ArticleItemInput item) -> contentRepoPersistenceService.createItem(item, ingestion))
+        .map((ArticleItemInput item) -> objectStorageService.createItem(item, ingestion))
         .collect(Collectors.toList());
     for (ArticleItem item : items) {
       hibernateTemplate.save(item);
     }
 
-    Collection<ArticleFile> files = contentRepoPersistenceService.persistAncillaryFiles(articlePackage,
+    Collection<ArticleFile> files = objectStorageService.persistAncillaryFiles(articlePackage,
         ingestion);
     for (ArticleFile file : files) {
       hibernateTemplate.save(file);
